requirements.txt
flask>=3.0.0
python-docx>=1.1.0
gunicorn>=22.0.0
